P4ADEV-2071-create-api-publish-debt-position
