+++ src/main/java/it/gov/pagopa/pu/debtpositions/service/statusalign/DebtPositionHierarchyStatusAlignerService.java
@@ -1,7 +1,6 @@
 package it.gov.pagopa.pu.debtpositions.service.statusalign;
 
 import it.gov.pagopa.pu.debtpositions.dto.BaseDebtPosition;
-import it.gov.pagopa.pu.debtpositions.dto.WfExecutionParameters;
 import it.gov.pagopa.pu.debtpositions.dto.generated.DebtPositionDTO;
 import it.gov.pagopa.pu.debtpositions.dto.generated.SyncStatusUpdateRequestDTO;
 import it.gov.pagopa.pu.debtpositions.dto.generated.TransferReportedRequest;
@@ -14,21 +13,18 @@ public interface DebtPositionHierarchyStatusAlignerService {
 
   /**
    * It will set the installment as REPORTED and then invoking sync debtPosition in order to publish the event
+   *
    * @return The DebtPosition updated and workflowId started
    */
   Pair<DebtPositionDTO, WorkflowCreatedDTO> notifyReportedTransferId(Long transferId, TransferReportedRequest transferReportedRequest, String accessToken);
 
   /**
    * It will verify installment expiration and then invoking sync debtPosition in order to publish the event and re-schedule expiration check
+   *
    * @return The DebtPosition updated and workflowId started
    */
   Pair<DebtPositionDTO, WorkflowCreatedDTO> checkAndUpdateInstallmentExpiration(Long debtPositionId, String accessToken);
 
   void alignHierarchyStatus(BaseDebtPosition debtPosition);
 
-  /**
-   * It will verify if debt position is draft and then invoking sync debtPosition in order to publish the debt position
-   * @return The DebtPosition updated and workflowId started
-   */
-  Pair<DebtPositionDTO, WorkflowCreatedDTO> publishDebtPosition(Long debtPositionId, WfExecutionParameters wfExecutionParameters, String accessToken, String operatorExternalUserId);
 }
